--- rfs/models/CNN/README.md
@@ -1,1 +1,0 @@
-# Convolutional Neural Network
