--- rfs/models/CNN/train.py
@@ -1,48 +1,0 @@
-import optax
-import torch
-import equinox as eqx
-
-from CNN import CNN
-from loss import loss_fn
-from eval import evaluate
-from jaxtyping import PyTree, Float, Int, Array
-
-
-def train(
-    model: CNN,
-    trainloader: torch.utils.data.DataLoader,
-    testloader: torch.utils.data.DataLoader,
-    optim: optax.GradientTransformation,
-    steps: int,
-    print_every: int,
-) -> CNN:
-    opt_state = optim.init(eqx.filter(model, eqx.is_array))
-
-    @eqx.filter_jit
-    def make_step(
-        model: CNN,
-        opt_state: PyTree,
-        x: Float[Array, "batch 1 28 28"],
-        y: Int[Array, " batch"],
-    ):
-        loss_value, grads = eqx.filter_value_and_grad(loss_fn)(model, x, y)
-        updates, opt_state = optim.update(grads, opt_state, model)
-        model = eqx.apply_updates(model, updates)
-        return model, opt_state, loss_value
-
-    def infinite_trainloader():
-        while True:
-            yield from trainloader
-
-    for step, (x, y) in zip(range(steps), infinite_trainloader()):
-        x = x.numpy()
-        y = y.numpy()
-        model, opt_state, train_loss = make_step(model, opt_state, x, y)
-        if (step % print_every) == 0 or (step == steps - 1):
-            test_loss, test_accuracy = evaluate(model, testloader)
-            print(
-                f"{step=}, train_loss={train_loss.item()}, "
-                f"test_loss={test_loss.item()}, test_accuracy={
-                    test_accuracy.item()}"
-            )
-    return model
