--- rfs/models/DDPM/README.md
@@ -1,2 +1,0 @@
-# Diffusion Models
-Implementing diffusion models in Python ([](https://arxiv.org/pdf/2006.11239))
