--- rfs/models/GAN/README.md
@@ -1,4 +1,0 @@
-# Generative adversarial networks
-
-### Acknowledgements
-The code is based on the [paper](https://arxiv.org/abs/1406.2661)
