--- rfs/models/GAN/main.py
@@ -1,9 +1,0 @@
-from training import train
-
-
-def main():
-    train()
-
-
-if __name__ == "__main__":
-    main()
