--- rfs/models/GAN/parameters.py
@@ -1,8 +1,0 @@
-class Params:
-    epochs = 50
-    batch_size = 128
-    learning_rate = 0.0002
-    latent_dim = 100  # size of the noise vector for generator
-    image_dim = 784  # for MNIST
-    num_channels = 1  # grayscale for MNIST
-    num_classes = 10  # for MNIST
