--- rfs/models/RNN/README.md
@@ -1,6 +1,0 @@
-# RNN
-
-### Recurrent Neural Networks
-
-### Acknowledgements
-The code is inspired by the paper "Learning Long-Term Dependencies with Gradient Descent is Difficult" by Yoshua Bengio et al. (1994).
