--- rfs/models/RNN/data_loader.py
@@ -1,35 +1,0 @@
-import math
-import numpy as np
-import jax.numpy as jnp
-import jax.random as jrandom
-
-
-def dataloader(arrays, batch_size):
-    dataset_size = arrays[0].shape[0]
-    assert all(array.shape[0] == dataset_size for array in arrays)
-    indices = np.arange(dataset_size)
-    while True:
-        perm = np.random.permutation(indices)
-        start = 0
-        end = batch_size
-        while end <= dataset_size:
-            batch_perm = perm[start:end]
-            yield tuple(array[batch_perm] for array in arrays)
-            start = end
-            end = start + batch_size
-
-
-def get_data(dataset_size, *, key):
-    t = jnp.linspace(0, 2 * math.pi, 16)
-    offset = jrandom.uniform(key, (dataset_size, 1),
-                             minval=0, maxval=2 * math.pi)
-    x1 = jnp.sin(t + offset) / (1 + t)
-    x2 = jnp.cos(t + offset) / (1 + t)
-    y = jnp.ones((dataset_size, 1))
-
-    half_dataset_size = dataset_size // 2
-    x1 = x1.at[:half_dataset_size].multiply(-1)
-    y = y.at[:half_dataset_size].set(0)
-    x = jnp.stack([x1, x2], axis=-1)
-
-    return x, y
