--- rfs/models/RNN/main.py
@@ -1,52 +1,0 @@
-import jax
-import optax
-import equinox as eqx
-import jax.numpy as jnp
-import jax.random as jrandom
-
-from data_loader import get_data, dataloader
-from model import RNN
-
-
-def main(
-    dataset_size=10000,
-    batch_size=32,
-    learning_rate=3e-3,
-    steps=200,
-    hidden_size=16,
-    depth=1,
-    seed=5678,
-):
-    data_key, model_key = jrandom.split(jrandom.PRNGKey(seed), 2)
-    xs, ys = get_data(dataset_size, key=data_key)
-    iter_data = dataloader((xs, ys), batch_size)
-
-    model = RNN(in_size=2, out_size=1, hidden_size=hidden_size, key=model_key)
-
-    @eqx.filter_value_and_grad
-    def compute_loss(model, x, y):
-        pred_y = jax.vmap(model)(x)
-        return -jnp.mean(y * jnp.log(pred_y) + (1 - y) * jnp.log(1 - pred_y))
-
-    @eqx.filter_jit
-    def make_step(model, x, y, opt_state):
-        loss, grads = compute_loss(model, x, y)
-        updates, opt_state = optim.update(grads, opt_state)
-        model = eqx.apply_updates(model, updates)
-        return loss, model, opt_state
-
-    optim = optax.adam(learning_rate)
-    opt_state = optim.init(model)
-    for step, (x, y) in zip(range(steps), iter_data):
-        loss, model, opt_state = make_step(model, x, y, opt_state)
-        loss = loss.item()
-        print(f"step={step}, loss={loss}")
-
-    pred_ys = jax.vmap(model)(xs)
-    num_correct = jnp.sum((pred_ys > 0.5) == ys)
-    final_accuracy = (num_correct / dataset_size).item()
-    print(f"final_accuracy={final_accuracy}")
-
-
-if __name__ == "__main__":
-    main()
