--- rfs/models/RNN/model.py
@@ -1,29 +1,0 @@
-import jax
-import jax.lax as lax
-import equinox as eqx
-import jax.numpy as jnp
-import jax.random as jrandom
-
-
-class RNN(eqx.Module):
-    hidden_size: int
-    cell: eqx.Module
-    linear: eqx.nn.Linear
-    bias: jax.Array
-
-    def __init__(self, in_size, out_size, hidden_size, *, key):
-        ckey, lkey = jrandom.split(key)
-        self.hidden_size = hidden_size
-        self.cell = eqx.nn.GRUCell(in_size, hidden_size, key=ckey)
-        self.linear = eqx.nn.Linear(
-            hidden_size, out_size, use_bias=False, key=lkey)
-        self.bias = jnp.zeros(out_size)
-
-    def __call__(self, input):
-        hidden = jnp.zeros((self.hidden_size,))
-
-        def f(carry, inp):
-            return self.cell(inp, carry), None
-
-        out, _ = lax.scan(f, hidden, input)
-        return jax.nn.sigmoid(self.linear(out) + self.bias)
