--- rfs/models/Unet/README.md
@@ -1,2 +1,0 @@
-# U-net
-Implementation of [U-net(https://arxiv.org/pdf/1505.04597)]
