--- rfs/models/VQ-GAN/README.md
@@ -1,3 +1,0 @@
-# VQ-GAN
-
-Implementation of [VQ-GAN](https://arxiv.org/abs/2012.09841)
