--- rfs/models/ViT/README.md
@@ -1,4 +1,0 @@
-# ViT
-
-### Vision Transformers
-Implementation of Vision Transformers (ViT) [paper](https://arxiv.org/abs/2010.11929).
