--- rfs/models/ViT/main.py
@@ -1,11 +1,0 @@
-from training import train
-from testing import test
-
-
-def main():
-    train()
-    test()
-
-
-if __name__ == "__main__":
-    main()
