--- rfs/models/backprop/README.md
@@ -1,1 +1,0 @@
-# Backpropagation
