+++ rfs/models/diffusion.py
@@ -2,6 +2,68 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
+from tqdm import tqdm
+
+
+class Diffusion:
+    def __init__(self, num_noise_steps=2, beta_start=1e-4, beta_end=2*1e-2, img_size=64):
+        self.num_noise_steps = num_noise_steps
+        self.beta_start = beta_start
+        self.beta_end = beta_end
+        self.img_size = img_size
+        self.device = torch.device(
+            'cuda' if torch.cuda.is_available() else 'cpu')
+
+        self.beta = self.prepare_noise_schedule().to(self.device)
+        self.alpha = 1 - self.beta
+        self.alpha_hat = torch.cumprod(self.alpha, dim=0)
+
+    def prepare_noise_schedule(self):
+        return torch.linspace(self.beta_start, self.beta_end, self.num_noise_steps)
+
+    def noise_images(self, x, t):
+        sqrt_alpha_hat = torch.sqrt(self.alpha_hat[t])[:, None, None, None]
+        sqrt_one_minus_alpha_hat = torch.sqrt(
+            1 - self.alpha_hat[t])[:, None, None, None]
+        noise = torch.randn_like(x)
+
+        return sqrt_alpha_hat*x + sqrt_one_minus_alpha_hat*noise, noise
+
+    def sample_timestemps(self, n):
+        return torch.randint(low=1, high=self.num_noise_steps, size=(n,))
+
+    def sample(self, model, n, labels, cfg_scales=3):
+        model.eval()
+
+        with torch.no_grad():
+            x = torch.randn(
+                (n, 3, self.img_size, self.img_size)).to(self.device)
+
+            for i in tqdm(reversed(range(1, self.num_noise_steps)), position=0):
+                t = (torch.ones(n) * i).long().to(self.device)
+                predicted_noise = model(x, t, labels)
+                if cfg_scales > 0:
+                    unconditional = model(x, t, None)
+                    predicted_noise = torch.lerp(
+                        unconditional, predicted_noise, cfg_scales)
+                alpha = self.alpha[t][:, None, None, None]
+                alpha_hat = self.alpha_hat[t][:, None, None, None]
+                beta = self.beta[t][:, None, None, None]
+
+                if i > 1:
+                    noise = torch.rand_like(x)
+                else:
+                    noise = torch.zeros_like(x)
+
+                x = 1 / torch.sqrt(alpha) * (x - ((1 - alpha) / (torch.sqrt(1 - alpha_hat)))
+                                             * predicted_noise) + torch.sqrt(beta) * noise
+
+                model.train()
+                x = (x.clamp(-1, 1) + 1) / 2
+                x = (x * 255).type(torch.uint8)
+
+                return x
+
 
 class EMA:
     def __init__(self, beta=0.99):
