--- rfs/models/fastKAN/README.md
@@ -1,2 +1,0 @@
-# Fast KAN
-Implementation of [this](https://arxiv.org/abs/2405.06721)
